requirements.txt
asgiref==3.5.2
Django==4.0.6
django-appconf==1.0.5
django-imagekit==4.1.0
django-js-asset==2.0.0
django-mathfilters==1.0.0
django-mptt==0.13.4
django-tinymce==3.5.0
django-treebeard==4.5.1
djangorestframework==3.13.1
pilkit==2.0
Pillow==9.2.0
psycopg2-binary==2.9.3
pytz==2022.2.1
six==1.16.0
sqlparse==0.4.2
